apps/runner/pyproject.toml
# sibyl-runner - Distributed agent execution daemon for Sibyl
# Connects to Sibyl Core, receives tasks, executes agents

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "../../VERSION"
pattern = "(?P<version>.+)"

[project]
name = "sibyl-runner"
dynamic = ["version"]
description = "Distributed agent runner daemon for Sibyl"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.13"
authors = [{ name = "Stefanie Jane", email = "stef@hyperbliss.tech" }]
keywords = ["ai-agents", "distributed", "runner", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

dependencies = [
    # CLI Framework
    "typer>=0.20.0",
    "rich>=13.0",
    # WebSocket Client
    "websockets>=15.0",
    # HTTP Client
    "httpx>=0.27",
    # Auth
    "pyjwt>=2.10",
    # Config
    "pyyaml>=6.0",
    # Async utilities
    "anyio>=4.7",
    # Shared library
    "sibyl-core",
]

# =============================================================================
# Development Dependencies (uv)
# =============================================================================

[dependency-groups]
dev = [
    "pyright",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
sibyl-runner = "sibyl_runner:main"

[project.urls]
Homepage = "https://github.com/hyperb1iss/sibyl"
Repository = "https://github.com/hyperb1iss/sibyl"

[tool.hatch.build.targets.wheel]
packages = ["src/sibyl_runner"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests", "/README.md"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF", "ASYNC"]
ignore = [
    "E501",  # Line length (formatter handles)
    "B008",  # Function call in defaults (standard Typer pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["sibyl_runner", "sibyl_core"]

[tool.pyright]
include = ["src"]
pythonVersion = "3.13"
typeCheckingMode = "standard"

# =============================================================================
# uv - Package Sources
# =============================================================================

[tool.uv.sources]
sibyl-core = { workspace = true }
